+++ src/lib/djerba/util/ini_fields.py
@@ -2,26 +2,24 @@
 
 # section names
 INPUTS = 'inputs'
-SEG = 'seg'
 SETTINGS = 'settings'
 DISCOVERED = 'discovered'
 
 # parameter names
-AMPL = 'ampl'
 ANALYSIS_UNIT = 'analysis_unit'
 BED_PATH = 'bed_path'
 DATA_DIR = 'data_dir'
 ENSCON = 'enscon'
 ENTCON = 'entcon'
-GAIN = 'gain'
-GAMMA = 'gamma'
 GENE_BED = 'genebed'
 GENE_LIST = 'genelist'
 GENOMIC_SUMMARY = 'genomic_summary'
 GEP_FILE = 'gepfile'
 GEP_REFERENCE = 'gep_reference'
-HTZD = 'htzd'
-HMZD = 'hmzd'
+LOG_R_AMPL = 'ampl'
+LOG_R_GAIN = 'gain'
+LOG_R_HTZD = 'htzd'
+LOG_R_HMZD = 'hmzd'
 MAF_FILE = 'maf_file'
 MAVIS_FILE = 'mavis_file'
 MEAN_COVERAGE = 'mean_coverage'
@@ -34,11 +32,17 @@ ONCOTREE_CODE = 'oncotree_code' # was CANCER_TYPE_DETAILED
 PATIENT = 'patient'
 PATIENT_ID = 'patientid'
 PCT_V7_ABOVE_80X = 'pct_v7_above_80x'
+PLOIDY = 'ploidy'
 PROVENANCE = 'provenance'
+PURITY = 'purity'
 REPORT_VERSION = 'report_version'
 SAMPLE_ANATOMICAL_SITE = 'sample_anatomical_site'
 SAMPLE_TYPE = 'sample_type'
 SEQUENZA_FILE = 'sequenza_file'
+SEQUENZA_GAMMA = 'sequenza_gamma'
+SEQUENZA_REVIEWER_1 = 'sequenza_reviewer_1'
+SEQUENZA_REVIEWER_2 = 'sequenza_reviewer_2'
+SEQUENZA_SOLUTION = 'sequenza_solution'
 SEX = 'sex'
 STUDY_ID = 'studyid'
 TCGA_CODE = 'tcgacode'
@@ -54,7 +58,10 @@ SCHEMA = {
         DATA_DIR,
         ENSCON,
         ENTCON,
-        GAMMA,
+        LOG_R_AMPL,
+        LOG_R_GAIN,
+        LOG_R_HMZD,
+        LOG_R_HTZD,
         GENE_BED,
         GENE_LIST,
         GENOMIC_SUMMARY,
@@ -66,7 +73,11 @@ SCHEMA = {
         ONCO_LIST,
         ONCOTREE_DATA,
         PATIENT_ID,
+        PLOIDY,
+        PURITY,
+        SEQUENZA_GAMMA,
         SEQUENZA_FILE,
+        SEQUENZA_SOLUTION,
         TMBCOMP,
         TUMOUR_ID
     ],
@@ -78,15 +89,11 @@ SCHEMA = {
         REPORT_VERSION,
         SAMPLE_ANATOMICAL_SITE,
         SAMPLE_TYPE,
+        SEQUENZA_REVIEWER_1,
+        SEQUENZA_REVIEWER_2,
         SEX,
         STUDY_ID,
         TCGA_CODE
-    ],
-    SEG: [
-        GAIN,
-        AMPL,
-        HTZD,
-        HMZD
     ],
     SETTINGS: [
         BED_PATH,