+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - GCGI-186: Setup mode in main script
 - GCGI-187: Fail and target-coverage options for HTML generation
 - GCGI-189: Script option to locate Sequenza results in file provenance
+- GCGI-193: Sequenza configuration/metadata fixes
 
 ## v0.0.6: 2021-08-19